tradescrubber/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Data sources
yfinance>=0.2.18
requests>=2.31.0

# Technical analysis
ta>=0.10.2

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# UI frameworks
nicegui>=1.4.0
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional API providers
# polygon-api-client>=1.12.0
# alpaca-py>=0.20.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
